Delete find_iris_spectra_moss.pro
Purpose of this file is to change the order of loops (AIA -> IRIS SG scan -> IRIS SG). However, it turned out that even though mosses are defined by AIA 193 A data, it is more efficient to use IRIS SG loop firstly because our goal is to find spectra observed by the IRIS. To check the existence of IRIS SG data is more efficient.